product card update
